+++ Kernel/System/Ticket/Event/NotificationEvent.pm
@@ -912,6 +912,9 @@ sub _SendRecipientNotification {
 
     my $TransportObject = $Param{TransportObject};
 
+    # get attachments
+    my @Attachments = @{ $Param{Attachments} };
+
     # send notification to each recipient
     my $Success = $TransportObject->SendNotification(
         TicketID              => $Param{TicketID},
@@ -920,7 +923,7 @@ sub _SendRecipientNotification {
         CustomerMessageParams => $Param{CustomerMessageParams},
         Recipient             => $Param{Recipient},
         Event                 => $Param{Event},
-        Attachments           => $Param{Attachments},
+        Attachments           => \@Attachments,
     );
 
     return if !$Success;